+++ spec/unit/installer/user_project_integrator/target_integrator_spec.rb
@@ -173,6 +173,25 @@ module Pod
           phase.nil?.should == false
         end
 
+        it 'does not add an embed frameworks build phase if the target to integrate is a messages extension for an iOS app' do
+          @pod_bundle.stubs(:requires_frameworks? => true)
+          target = @target_integrator.send(:native_targets).first
+          target.stubs(:symbol_type).returns(:messages_extension)
+          @target_integrator.integrate!
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.nil?.should == true
+        end
+
+        it 'adds an embed frameworks build phase if the target to integrate is a messages extension for a messages application' do
+          @pod_bundle.stubs(:requires_frameworks? => true)
+          @pod_bundle.stubs(:requires_host_target? => false) # Messages extensions for messages applications do not require a host target
+          target = @target_integrator.send(:native_targets).first
+          target.stubs(:symbol_type).returns(:messages_extension)
+          @target_integrator.integrate!
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.nil?.should == false
+        end
+
         it 'adds an embed frameworks build phase if the target to integrate is a UI Test bundle' do
           @pod_bundle.stubs(:requires_frameworks? => true)
           target = @target_integrator.send(:native_targets).first
@@ -224,6 +243,31 @@ module Pod
           @target_integrator.integrate!
           phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
           phase.nil?.should == true
+        end
+
+        it 'removes embed frameworks build phases from messages extension targets that are used in an iOS app' do
+          @pod_bundle.stubs(:requires_frameworks? => true)
+          @target_integrator.integrate!
+          target = @target_integrator.send(:native_targets).first
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.nil?.should == false
+          target.stubs(:symbol_type).returns(:messages_extension)
+          @target_integrator.integrate!
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.nil?.should == true
+        end
+
+        it 'does not remove embed frameworks build phases from messages extension targets that are used in a messages app' do
+          @pod_bundle.stubs(:requires_frameworks? => true)
+          @target_integrator.integrate!
+          target = @target_integrator.send(:native_targets).first
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.nil?.should == false
+          target.stubs(:symbol_type).returns(:messages_extension)
+          @pod_bundle.stubs(:requires_host_target? => false) # Messages extensions for messages applications do not require a host target
+          @target_integrator.integrate!
+          phase = target.shell_script_build_phases.find { |bp| bp.name == @embed_framework_phase_name }
+          phase.nil?.should == false
         end
 
         it 'removes embed frameworks build phases from framework targets' do